# [wifi] Replace std::stable_sort with insertion sort to save 2.4KB flash